Implemented below features 1. Present value will be displayed in decimals considering Half-Day working 2. Half day (HD-0.5P), National Holiday present(HP-2P) and Festival leave present(FP-2P) attendance codes added into the system with their implications in Payroll Calculations 3. Corrections is master page, logout functionality is optimized.

+++ WebApplication1/bussiness/production/gen_agl_F17.aspx.cs
@@ -26,7 +26,7 @@ namespace WebApplication1.bussiness.production
         public static string date1 = "";
         public static string date2 = "";
         public static Int32 CalWorkingDays = 0;
-        public static Int32 TotalPresents = 0;
+        public static decimal TotalPresents = 0;
         public static decimal GorssBreaker = 0;
 
         protected void Page_Load(object sender, EventArgs e)
